Fixed issue with folders
I apologize for the confusion. I now realize that the issue could be caused by the trailing space in the folder name "PHP introduction screenshot Codecademy ". It's likely that the operating system is removing the trailing space when creating the folder, but the path still contains the space when attempting to create the file.