File Append Http
I need some help with this we really need to do this please show me how In the DownloadFileFromGoogleDrive method, you're creating a MemoryStream to download the file, then saving it to disk using FileMode.Create. This means that if the file already exists, the method will overwrite it. If you expect to download the same file multiple times, you could potentially use FileMode.Append instead to avoid overwriting the file every time. Additionally, you could potentially use HttpClient instead of Google.Apis.Download to download the file, as this may be faster for large files.